requirements.txt
pandas
google-generativeai
flask
